minor schema and db updates, added api routes to build and return matchSession and matchRecord dtos

+++ models/Swipe.js
@@ -24,7 +24,7 @@ Swipe.init(
             type: DataTypes.INTEGER,
         },
         distance: {
-            type: DataTypes.FLOAT,       
+            type: DataTypes.FLOAT(10, 2),       
         },
         message: {
             type: DataTypes.STRING,
